pyproject.toml
[tool.poetry]
name = "fitness-api"
version = "1.0.0"
description = "API for a fitness tracking web app"
authors = ["Aman Mughal <aman.mughal@amppa.dev>"]
readme = "README.md"
packages = [{include = "fitness_api"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "0.103.2"
pydantic = "*"
pydantic-settings = "2.0.3"
SQLAlchemy = "2.0.21"
uvicorn = "0.20.0"
psycopg2-binary = "2.9.7"
loguru = "0.7.2"
requests = "2.28.2"
python-jose = { version = "3.3.0", extras = ["cryptography"] }
passlib = { version = "1.7.4", extras = ["bcrypt"] }
python-multipart = "0.0.6"

[tool.poetry.group.dev.dependencies]
black = { version = "*", extras = ["jupyter"] }
ruff = "*"
pytest = "*"
pytest-order = "*"
mypy = "*"
pkginfo = "*"
wheel = "*"
setuptools = "*"
autopep8 = "*"
isort = "*"
flake8 = "*"
pep8-naming = "*"
pytest-asyncio = "*"
coverage = "*"
Sphinx = "*"
nbsphinx = "*"
furo = "*"
sphinx-inline-tabs = "*"
sphinx-copybutton = "*"
sphinx-gallery = "*"
sphinx-autodoc-typehints = "*"
sphinx-autobuild = "*"
matplotlib = "*"
seaborn = "*"
types-requests = "*"
types-PyYAML = "*"
types-python-dateutil = "*"
jupyterlab-code-formatter = "*"
jupyterlab-spellchecker = "*"
pyinstrument = "*"
httpx = "*"
pre-commit = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py311"]
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | node_modules
)/
'''


[tool.ruff]
select = ["E", "F"]
ignore = ["N817", "N806", "N803", "N802", "N811", "N814"]
exclude = [
  ".bzr",
  ".direnv",
  ".eggs",
  ".git",
  ".hg",
  ".mypy_cache",
  ".nox",
  ".pants.d",
  ".pytype",
  ".ruff_cache",
  ".svn",
  ".tox",
  ".venv",
  "__pypackages__",
  "__pycache__",
  "_build",
  "buck-out",
  "build",
  "dist",
  "node_modules",
  "venv",
]

line-length = 120

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

target-version = "py311"

[tool.ruff.mccabe]
max-complexity = 10
